+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - [MODOS] Amélioration des performances et de la précision de la recherche d'utilisateurs
 - [MODOS] Suppression automatique des titres de la bibliothèque Blinest des playlists associées avec notification Discord
 - [GAME] Refonte complète de la page de profil utilisateur
+- [CHAT] Réactions sur les messages